requirements.txt
cycler==0.11.0
fonttools==4.28.2
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
